--- vms/tests/mod.rs
@@ -1,45 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018-2019 Aion foundation.
- *
- *     This file is part of the aion network project.
- *
- *     The aion network project is free software: you can redistribute it
- *     and/or modify it under the terms of the GNU General Public License
- *     as published by the Free Software Foundation, either version 3 of
- *     the License, or any later version.
- *
- *     The aion network project is distributed in the hope that it will
- *     be useful, but WITHOUT ANY WARRANTY; without even the implied
- *     warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- *     See the GNU General Public License for more details.
- *
- *     You should have received a copy of the GNU General Public License
- *     along with the aion network project source files.
- *     If not, see <https://www.gnu.org/licenses/>.
- *
- ******************************************************************************/
-
-#[cfg(test)]
-mod bytes {
-    /*
-    fn merge(list: Vec<[u8]>) -> &[u8] {
-        let size: i32 = 0;
-        let count = list.len();
-    }
-    
-    #[test]
-    fn test_compilation() {
-    
-    }
-    */
-}
-
-#[cfg(test)]
-mod contract_test {
-    // extern crate fastvm;
-    // extern crate vms;
-
-    // use fastvm::vm::{Result, GasLeft, Ext, ActionParams, ActionValue};
-
-    // let params = ActionParams::default();
-}
